(Abstract) com_render, com_render_diffuse, sys_render
