+++ Abstract/components/com_render.ts
@@ -1,0 +1,7 @@
+import {RenderDiffuse} from "./com_render_diffuse.js";
+
+export type Render = RenderDiffuse;
+
+export const enum RenderKind {
+    Diffuse,
+}
