+++ Abstract/components/com_render_diffuse.ts
@@ -1,0 +1,58 @@
+import {Material, Mesh} from "../../common/material.js";
+import {Vec4} from "../../common/math.js";
+import {GL_ARRAY_BUFFER, GL_CW, GL_ELEMENT_ARRAY_BUFFER, GL_FLOAT} from "../../common/webgl.js";
+import {DiffuseLayout} from "../../materials/layout_diffuse.js";
+import {Entity, Game} from "../game.js";
+import {Has} from "../world.js";
+import {RenderKind} from "./com_render.js";
+
+export interface RenderDiffuse {
+    readonly Kind: RenderKind.Diffuse;
+    readonly Material: Material<DiffuseLayout>;
+    readonly Mesh: Mesh;
+    readonly FrontFace: GLenum;
+    readonly Vao: WebGLVertexArrayObject;
+    Color: Vec4;
+}
+
+let vaos: WeakMap<Mesh, WebGLVertexArrayObject> = new WeakMap();
+
+export function render_diffuse(material: Material<DiffuseLayout>, mesh: Mesh, color: Vec4) {
+    return (game: Game, entity: Entity) => {
+        if (!vaos.has(mesh)) {
+            // We only need to create the VAO once.
+            let vao = game.ExtVao.createVertexArrayOES()!;
+            game.ExtVao.bindVertexArrayOES(vao);
+
+            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.VertexBuffer);
+            game.Gl.enableVertexAttribArray(material.Locations.VertexPosition);
+            game.Gl.vertexAttribPointer(
+                material.Locations.VertexPosition,
+                3,
+                GL_FLOAT,
+                false,
+                0,
+                0
+            );
+
+            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.NormalBuffer);
+            game.Gl.enableVertexAttribArray(material.Locations.VertexNormal);
+            game.Gl.vertexAttribPointer(material.Locations.VertexNormal, 3, GL_FLOAT, false, 0, 0);
+
+            game.Gl.bindBuffer(GL_ELEMENT_ARRAY_BUFFER, mesh.IndexBuffer);
+
+            game.ExtVao.bindVertexArrayOES(null);
+            vaos.set(mesh, vao);
+        }
+
+        game.World.Signature[entity] |= Has.Render;
+        game.World.Render[entity] = {
+            Kind: RenderKind.Diffuse,
+            Material: material,
+            Mesh: mesh,
+            FrontFace: GL_CW,
+            Vao: vaos.get(mesh)!,
+            Color: color,
+        };
+    };
+}
