+++ Abstract/world.ts
@@ -1,16 +1,19 @@
 import {Camera} from "./components/com_camera.js";
 import {Light} from "./components/com_light.js";
+import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
     Camera,
     Light,
+    Render,
     Transform,
 }
 
 export const enum Has {
     Camera = 1 << Component.Camera,
     Light = 1 << Component.Light,
+    Render = 1 << Component.Render,
     Transform = 1 << Component.Transform,
 }
 
@@ -21,5 +24,6 @@ export interface World {
     // Component data
     Camera: Array<Camera>;
     Light: Array<Light>;
+    Render: Array<Render>;
     Transform: Array<Transform>;
 }
