+++ NewProject3D/components/com_render.ts
@@ -1,7 +1,1 @@
-import {RenderDiffuse} from "./com_render_diffuse.js";
-
-export type Render = RenderDiffuse;
-
-export const enum RenderKind {
-    Diffuse,
-}
+../../Abstract/components/com_render.ts
